[Move] a new dir for keyboard related code in browser
