+++ src/code/browser/service/keyboard/keyboardScreenCastService.ts
@@ -3,7 +3,7 @@ import { IDisposable } from "src/base/common/dispose";
 import { DomUtility } from "src/base/common/dom";
 import { IStandardKeyboardEvent, Keyboard } from "src/base/common/keyboard";
 import { IntervalTimer } from "src/base/common/util/timer";
-import { IKeyboardService } from "src/code/browser/service/keyboardService";
+import { IKeyboardService } from "src/code/browser/service/keyboard/keyboardService";
 import { createDecorator } from "src/code/platform/instantiation/common/decorator";
 
 export const IKeyboardScreenCastService = createDecorator<IKeyboardScreenCastService>('keyboard-screencast-service');
