+++ src/code/browser/service/keyboard/keyboardService.ts
@@ -3,8 +3,6 @@ import { EventType } from "src/base/common/dom";
 import { createDecorator } from "src/code/platform/instantiation/common/decorator";
 import { DomEmitter, Event, Register } from "src/base/common/event";
 import { createStandardKeyboardEvent, IStandardKeyboardEvent } from "src/base/common/keyboard";
-import { registerSingleton } from "src/code/platform/instantiation/common/serviceCollection";
-import { ServiceDescriptor } from "src/code/platform/instantiation/common/descriptor";
 
 export const IKeyboardService = createDecorator<IKeyboardService>('keyboard-service');
 
@@ -52,5 +50,3 @@ export class keyboardService implements IDisposable, IKeyboardService {
     }
 
 }
-
-registerSingleton(IKeyboardService, new ServiceDescriptor(keyboardService));
